requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-crispy-forms==2.1
django-multiselectfield==0.1.12
pillow==10.2.0
psycopg==3.1.17
psycopg-binary==3.1.17
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
